Created a CartIcon, CartDropdown, and ProductCard React component. The CartIcon will function as the button that the user can press to then have the CartDropdown displayed. The CartDropdown will have a list of all of the items they have added to their cart. The ProductCard component will display an image for the product along with its price, name, and a button that the user can click to add that product to their shopping cart.
Created a Shop component that will format and render all of the
ProductCards in a neat fashion.

Created a ProductContext, and CartDropdownContext. The ProductContext
will keep track of the products that will be sold on the web app. The
CartDropdownContext will keep track of whether the CartDropdown
component should be displayed on the screen.

Brian M.